tokio_websockets

Module client

Implementation of a WebSocket client.

This can be used in three ways:

By letting the library connect to a remote URI and performing a HTTP/1.1 Upgrade handshake, via Builder::connect

By letting the library perform a HTTP/1.1 Upgrade handshake on an established stream, via Builder::connect_on

By performing the handshake yourself and then using Builder::take_over to let it take over a WebSocket stream

Structs

Builder

Builder for WebSocket client connections.

Constants

DISALLOWED_HEADERS

List of headers added by the client which will cause an error if added by the user: