tokio_websockets::proto

Struct Message

pub struct Message { /* private fields */ }

A WebSocket message. This is cheaply clonable and uses Payload as the payload storage underneath.

Received messages are always validated prior to dealing with them, so all the type casting methods are either almost or fully zero cost.

Implementations

impl Message

pub fn text<P: Into<Payload>>(payload: P) -> Self

Create a new text message. The payload contents must be valid UTF-8.

pub fn binary<P: Into<Payload>>(payload: P) -> Self

Create a new binary message.

pub fn close(code: Option<CloseCode>, reason: &str) -> Self

Create a new close message. If an non-empty reason is specified, a CloseCode must be specified for it to be included.

Panics

If the code is reserved so it cannot be sent.

If code is present and the reason exceeds 123 bytes, the protocol-imposed limit.

pub fn ping<P: Into<Payload>>(payload: P) -> Self

Create a new ping message.

Panics

If the payload exceeds 125 bytes, the protocol-imposed limit.

pub fn pong<P: Into<Payload>>(payload: P) -> Self

Create a new pong message.

Panics

If the payload exceeds 125 bytes, the protocol-imposed limit.

pub fn is_text(&self) -> bool

Whether the message is a text message.

pub fn is_binary(&self) -> bool

Whether the message is a binary message.

pub fn is_close(&self) -> bool

Whether the message is a close message.

pub fn is_ping(&self) -> bool

Whether the message is a ping message.

pub fn is_pong(&self) -> bool

Whether the message is a pong message.

pub fn into_payload(self) -> Payload

Returns the message payload and consumes the message, regardless of type.

pub fn as_payload(&self) -> &Payload

Returns a reference to the message payload, regardless of message type.

pub fn as_text(&self) -> Option<&str>

Returns a reference to the message payload as a string if it is a text message.

Panics

This method will panic when the message was created via Message::text with invalid UTF-8.

pub fn as_close(&self) -> Option<(CloseCode, &str)>

Returns the CloseCode and close reason if the message is a close message.